feat!: move all color-converter utils (`addon-editor` / `addon-doc` => `cdk`)
