+++ projects/addon-editor/utils/index.ts
@@ -3,14 +3,9 @@ export * from './get-element-point';
 export * from './get-gradient-data';
 export * from './get-mark-range';
 export * from './get-selected-content';
-export * from './hex-to-rgb';
-export * from './hsv-to-rgb';
 export * from './insert-html';
 export * from './insert-text';
 export * from './is-selection-in';
 export * from './legacy-converter';
-export * from './parse-color';
 export * from './parse-gradient';
-export * from './rgb-to-hex';
 export * from './to-gradient';
-export * from './tui-rgb-to-hsv';
