+++ projects/cdk/utils/color/hex-to-rgb.ts
@@ -1,11 +1,3 @@
-/**
- * TODO: duplicate, need fix it before v3
- * TODO: move another package for reuse between addon-editor and addon-doc
- */
-export function tuiRgbToHex(r: number, g: number, b: number): string {
-    return `#${[r, g, b].map(x => x.toString(16).padStart(2, `0`)).join(``)}`;
-}
-
 export function tuiHexToRgb(hex: string): [number, number, number] {
     const matches = hex
         .replace(`#`, ``)
