+++ projects/cdk/utils/color/package.json
@@ -1,0 +1,7 @@
+{
+    "ngPackage": {
+        "lib": {
+            "entryFile": "index.ts"
+        }
+    }
+}
