+++ projects/cdk/utils/color/rgb-to-hex.ts
@@ -1,0 +1,3 @@
+export function tuiRgbToHex(r: number, g: number, b: number): string {
+    return `#${[r, g, b].map(x => x.toString(16).padStart(2, `0`)).join(``)}`;
+}
