+++ projects/cdk/utils/index.ts
@@ -1,4 +1,5 @@
 export * from '@taiga-ui/cdk/utils/browser';
+export * from '@taiga-ui/cdk/utils/color';
 export * from '@taiga-ui/cdk/utils/dom';
 export * from '@taiga-ui/cdk/utils/focus';
 export * from '@taiga-ui/cdk/utils/format';
